refactor(tools): split DelegationTools into focused delegation/ modules
DelegationTools.ts mixed three concerns in 1219 LOC: subagent execution +
delivery lifecycle, the propose/approve handshake + agent-model resolution,
and the input-field schemas + attachment validation. Extract each internal
helper cluster into a src/tools/delegation/ sibling module; keep the two
tool-definition classes, their schemas, and the public API in DelegationTools.ts
as a thin facade (re-exporting rejectOversizedBibAttachments). Pure move: no
logic change, public API and every importer unchanged.
