+++ src/tools/delegation/proposalFlow.ts
@@ -1,0 +1,216 @@
+/**
+ * Proposal/approval handshake and agent/model resolution for delegation tools.
+ *
+ * If proposal bypass is active for a stream, launches immediately; otherwise
+ * waits for user approval via the proposal coordinator before executing.
+ */
+
+// Third-party imports
+import { nanoid } from 'nanoid';
+
+// Local imports - agent
+import { getVisibleAgent, getVisibleAgents } from '@agent/index/agentRegistry';
+import { currentSession } from '@agent/runtime/SessionHandle';
+import type { ProposalResult } from '@agent/runtime/AgentProposalCoordinator';
+
+// Local imports - model
+import { computeModelOptionsData } from '@model/computeModelOptions';
+
+// Local imports - tools
+import {
+  AgentCategory,
+  type WorkflowAgentProposal,
+  type ToolUseAgentProposal,
+  type StreamTabId,
+} from '@shared/schemas';
+import type { ToolResult } from '@shared/schemas/toolResult';
+import {
+  isApprovalBypassedForStream,
+  proposalApprovalState,
+} from '@tools/approval';
+import {
+  availableModelNamesFromOptions,
+  resolveDelegationModelFromAvailableNames,
+} from '@tools/delegationModelAvailability';
+
+// Local imports - delegation
+import { executeSubagent } from './subagentExecution';
+
+const DEFAULT_DELEGATION_REJECTION_FEEDBACK = [
+  'No feedback provided.',
+  'Do not retry the same or equivalent delegation unless the user explicitly asks for it;',
+  'continue directly with available context, or ask the user a clarifying question.',
+].join(' ');
+
+/** Format an agent list for tool descriptions. */
+export function formatAgentList(
+  agents: { name: string; description?: string; tools?: string[] }[],
+): string {
+  return agents
+    .map((agent) => {
+      const desc = agent.description || 'No description';
+      const toolsSuffix = agent.tools?.length
+        ? `\n  Tools: ${agent.tools.join(', ')}`
+        : '';
+      return `- ${agent.name}: ${desc}${toolsSuffix}`;
+    })
+    .join('\n');
+}
+
+export async function resolveAvailableDelegationModel(input: {
+  readonly requestedModel?: string | null;
+  readonly parentModel?: string | null;
+}): Promise<string> {
+  const modelOptions = await computeModelOptionsData();
+  return resolveDelegationModelFromAvailableNames({
+    ...input,
+    availableModels: availableModelNamesFromOptions(modelOptions),
+  });
+}
+
+/** Return the visible current agent, or throw with the current visible list. */
+export function requireVisibleAgent(
+  category: AgentCategory,
+  name: string,
+): { name: string } {
+  const agent = getVisibleAgent(category, name);
+  if (agent) return agent;
+  const available = getVisibleAgents(category)
+    .map((a) => a.name)
+    .join(', ');
+  throw new Error(
+    `Unknown ${category} agent '${name}'. Available: ${available}`,
+  );
+}
+
+/** Build a concise summary of proposal parameters for rejection echo. */
+function summarizeProposal(
+  proposal: WorkflowAgentProposal | ToolUseAgentProposal,
+): string {
+  const parts = [`Agent: ${proposal.agent}`, `Model: ${proposal.model}`];
+  if ('inputFiles' in proposal && proposal.inputFiles?.[0]) {
+    parts.push(`File: ${proposal.inputFiles[0]}`);
+  }
+  if (proposal.memories.length > 0) {
+    parts.push(`Memories: ${proposal.memories.join(', ')}`);
+  }
+  const instrPreview =
+    proposal.instruction.length > 120
+      ? `${proposal.instruction.slice(0, 117)}...`
+      : proposal.instruction;
+  parts.push(`Instruction: "${instrPreview}"`);
+  return parts.join(', ');
+}
+
+/** Convert proposal result to ToolResult. Returns null if approved. */
+function proposalResultToToolResult(
+  result: ProposalResult,
+  agentName: string,
+  proposal: WorkflowAgentProposal | ToolUseAgentProposal,
+): ToolResult | null {
+  const echo = summarizeProposal(proposal);
+
+  switch (result.action) {
+    case 'reject': {
+      const feedback = result.feedback?.trim();
+      const feedbackLine = feedback
+        ? `\nUser feedback: ${feedback}`
+        : `\n${DEFAULT_DELEGATION_REJECTION_FEEDBACK}`;
+      return {
+        summary: `User rejected delegation to '${agentName}'`,
+        output: `Delegation to '${agentName}' was rejected.\nYour delegation was: ${echo}${feedbackLine}`,
+        isError: true,
+      };
+    }
+    case 'timeout':
+      return {
+        summary: `Delegation to '${agentName}' timed out`,
+        output: `Delegation to '${agentName}' timed out waiting for user approval.\nYour delegation was: ${echo}`,
+        isError: true,
+      };
+    case 'setup':
+      return {
+        summary: `User opened '${agentName}' for editing`,
+        output: `Delegation opened for editing. The user will run it manually when ready.\nYour delegation was: ${echo}`,
+      };
+    case 'approve':
+      return null;
+  }
+}
+
+/**
+ * Shared proposal-or-bypass flow used by both delegate_workflow and delegate_agent.
+ *
+ * If proposal bypass is active for this stream, skips the proposal and launches immediately.
+ * Otherwise, waits for user approval via the proposal coordinator.
+ */
+export async function proposeAndExecute(
+  proposal: WorkflowAgentProposal | ToolUseAgentProposal,
+  agentName: string,
+  streamId: StreamTabId,
+): Promise<ToolResult> {
+  if (proposalApprovalState.isBypassed(streamId)) {
+    return executeSubagent(proposal, agentName, streamId, {
+      enableYoloOnChild: true,
+      approvalMeta: { autoApproved: true },
+    });
+  }
+
+  const proposalId = nanoid();
+
+  const result = await currentSession().coordinators.waitForProposal(streamId, {
+    proposalId,
+    proposal,
+  });
+
+  const nonApproveResult = proposalResultToToolResult(
+    result,
+    agentName,
+    proposal,
+  );
+  if (nonApproveResult) return nonApproveResult;
+
+  // At this point result.action === 'approve' (all other cases returned above).
+  if (result.action !== 'approve') {
+    throw new Error(`Unexpected non-approve proposal result: ${result.action}`);
+  }
+  const modelOverride = result.model;
+  const agentOverride =
+    result.agent && result.agent !== proposal.agent ? result.agent : undefined;
+  const resolvedAgentOverride = agentOverride
+    ? getVisibleAgent(proposal.agentCategory, agentOverride)?.name
+    : undefined;
+
+  // Re-validate against the current registry — between proposal display and
+  // approval the agent may have been removed/renamed, or the approval could
+  // carry a malformed value. Fail fast so the orchestrator sees the problem
+  // synchronously instead of after an async launch.
+  if (agentOverride && !resolvedAgentOverride) {
+    return {
+      summary: `Approved agent override '${agentOverride}' is not available`,
+      output: `Cannot launch '${agentOverride}': it is not currently a visible ${proposal.agentCategory} agent (removed, renamed, or disabled since the proposal was shown). Re-propose the delegation.`,
+      isError: true,
+    };
+  }
+
+  const effective = {
+    ...proposal,
+    ...(modelOverride && { model: modelOverride }),
+    ...(resolvedAgentOverride && { agent: resolvedAgentOverride }),
+  };
+  const effectiveAgentName = resolvedAgentOverride ?? agentName;
+  return executeSubagent(effective, effectiveAgentName, streamId, {
+    enableYoloOnChild: isApprovalBypassedForStream(streamId),
+    approvalMeta: {
+      autoApproved: false,
+      ...(modelOverride && {
+        modelOverride,
+        requestedModel: proposal.model,
+      }),
+      ...(agentOverride && {
+        agentOverride,
+        requestedAgent: proposal.agent,
+      }),
+    },
+  });
+}
